[Improve?] Removed the shared transition from the image viewer back to the list of images since it is a bit off still and I couldn't figure out where it was coming from.
With this, since there was no transition in the activity, it called the Fragment.onStart method again, which is where I moved the data reloading too. To solve this, I added a flag in the shared preferences for whether or not we are coming from the photo viewer. If we are, then don't reload the data, since it isn't necessary.